Don't create git repo if child of an existing one
Before, generating a react project into an existing Rails one, the user
would end up with two nested git repositories: the Rails repo and
another one inside client/. If the user forgets to delete the client/
one, git becomes very upset. This makes setting up Rails+React more
complicated than it needs to be.

Now we detect if a .git directory already exists in the parent. If so,
we skip the `git init` step.

Update the Rails+React docs to reflect this change.