Install deps in CircleCI cron workflow
The cron workflow is currently failing for other spraygun repos because
of missing dependencies:

https://circleci.com/gh/carbonfive/spraygun-express/1101

We haven't seen this problem yet in spraygun-react-ts, but it seems like
it could happen.

Fix by ensuring that the `install_deps` job is run in the cron workflow.

I verified the CircleCI syntax with:

```
$ circleci config validate
Config file at .circleci/config.yml is valid.
```