Change lint config files to .js; document BEM rule
Before, our eslint and stylelint rules were in .json format, which
doesn't have a nice way of supporting comments.

Fix by switch in to .js format and add comments for the particularly
tricky part of the stylelint config where we use a big regex to enforce
BEM class names.

Also:

- Simplify the BEM regexp by removing the `u-*` exception (this was
  intended to support utility class names, but has not really been
  socialized).
- Add a friendly error message that is shown when the BEM and other
  naming rules are violated, instead of just complaining the regexp
  wasn't matched.
- Fix a quoting issue in the `lint:stylelint` task that caused stylelint
  to not actually check any .scss files!