Update Rails recommendations; remove Redux
Rails has a good, reliable integration with React via webpacker. It
comes with Rails out of the box, and has the benefit of convention over
configuration.

This commit updates our official recommendation for Rails+React to use
the built-in webpacker approach. I've link to the documentation for this
that already exists in the raygun README.

Next, this commit removes our Redux recommendations from the README.
Redux may still have its place on some projects, but I believe it
shouldn't be something that we reach for right away. The goal of
spraygun is to get people started with something quickly for a
greenfield app. There are simpler ways to manage state in React that
work great on small code bases and don't have the Redux learning curve.

Finally, update the instructions to require Node 16.